iot-analytic/requirements.txt
PyYAML == 5.4.1
